pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bcrypt>=5.0.0",
    "celery[redis]>=5.3.1",
    "dj-redis-panel>=0.6.0",
    "django>=5.2.7",
    "django-anymail[resend]>=14.0",
    "django-celery-beat>=2.8.1",
    "django-cors-headers>=4.9.0",
    "django-filter>=25.2",
    "django-json-widget>=2.1.0",
    "django-prometheus>=2.4.1",
    "django-redis>=6.0.0",
    "django-structlog[celery]>=9.1.1",
    "djangorestframework>=3.16.1",
    "djangorestframework-api-key==3.*",
    "djangorestframework-simplejwt>=5.5.1",
    "drf-spectacular>=0.28.0",
    "gunicorn>=24.1.1",
    "httpx>=0.28.1",
    "markdown>=3.9",
    "orjson>=3.11.5",
    "psycopg[binary,pool]>=3.2.10",
    "pydantic>=2.11.10",
    "pydantic-settings>=2.11.0",
    "redis[hiredis]>=6.4.0",
    "requests>=2.32.5",
    "whitenoise>=6.11.0",
]

[dependency-groups]
dev = [
    "coverage>=7.13.3",
    "django-stubs>=5.2.5",
    "djlint>=1.36.4",
    "model-bakery>=1.23.2",
    "pytest>=9.0.2",
    "pytest-django>=4.11.1",
    "ruff>=0.14.0",
    "ty>=0.0.11",
    "types-requests>=2.32.4.20250913",
]

[tool.uv]
package = false

[tool.django-stubs]
django_settings_module = "core.config.django.local"

[tool.ruff]
line-length = 120
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "B", "Q", "RUF100", "C90", "UP", "I", "DJ"]
flake8-quotes = { inline-quotes = "single", multiline-quotes = "double" }
mccabe = { max-complexity = 15 }

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint.isort]
combine-as-imports = true
section-order = ["future", "standard-library", "third-party", "first-party", "local-folder"]
known-first-party = ["backend"]


[tool.ruff.format]
quote-style = "single"

[tool.djlint]
profile = "django"
indent = 2
ignore = "H006,H030,H031,H021"
blank_line_after_tag = "load,extends"
format_css = true
format_js = false
max_line_length = 120
